fix (apparent) data loss due to V8 clock bug
When V8's clock mysteriously sticks (oh yes, it happens), and the stuck daemon
process is restarted, it may end up appending points out of order.

This eventually causes the points to get truncated because they are assumed
to be in sorted order. There is no actual data lost, but the graph gets cut
off.

This works around it by skipping over points that backtrack.

The LevelDB backend shouldn't need this fix because it doesn't do row coalescing.